common.cpp: created
Moved common parts from client.cpp and server.cpp
